Changed type for pitch bend.

+++ src/sfz.cpp
@@ -45,7 +45,7 @@ namespace sfz
 
 	bool
 	Region::OnKey(uint8_t chan, uint8_t key, uint8_t vel,
-		      uint8_t bend, uint8_t bpm, uint8_t chanaft, uint8_t polyaft,
+		      int bend, uint8_t bpm, uint8_t chanaft, uint8_t polyaft,
 		      uint8_t prog, float rand, trigger_t trig, uint8_t* cc,
 		      float timer, uint8_t seq, bool* sw, uint8_t last_sw_key, uint8_t prev_sw_key)
 	{
@@ -101,7 +101,7 @@ namespace sfz
 
 	bool
 	Region::OnControl(uint8_t chan, uint8_t cont, uint8_t val,
-			  uint8_t bend, uint8_t bpm, uint8_t chanaft, uint8_t polyaft,
+			  int bend, uint8_t bpm, uint8_t chanaft, uint8_t polyaft,
 			  uint8_t prog, float rand, trigger_t trig, uint8_t* cc,
 			  float timer, uint8_t seq, bool* sw, uint8_t last_sw_key, uint8_t prev_sw_key)
 	{
